requirements.txt
# flask
# gunicorn
openai
fastapi[all]
uvicorn[standard]
httpx
python-dotenv
azure-search-documents
line-bot-sdk
pymongo
python-binary-memcached
google-genai
sentence_transformers